requirements2.txt
asgiref==3.2.10
uWSGI>=2.0.18,<2.1
attrs==20.2.0
Automat==20.2.0
bcrypt==3.2.0
cached-property==1.5.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
constantly==15.1.0
cryptography==3.1.1
cssselect==1.1.0
distro==1.5.0
Django==3.1.2
djangorestframework==3.12.1
docker==4.3.1
docker-compose==1.27.4
dockerpty==0.4.1
docopt==0.6.2
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
itemadapter==0.1.1
itemloaders==1.0.3
jmespath==0.10.0
jsonschema==3.2.0
lxml==4.6.1
paramiko==2.7.2
parsel==1.6.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
PyNaCl==1.4.0
pyOpenSSL==19.1.0
pyrsistent==0.17.3
python-dotenv==0.14.0
python-scrapyd-api==2.1.2
pytz==2020.1
PyYAML==5.3.1
queuelib==1.5.0
requests==2.24.0
Scrapy==2.4.0
scrapyd==1.2.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
texttable==1.6.3
Twisted==20.3.0
urllib3==1.25.11
w3lib==1.22.0
websocket-client==0.57.0
zope.interface==5.1.2
